added part for non static methods
