+++ Runtime/Scripts/Core/ConsoleMethodInfo.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c5c682bddc564e541a62468e4a1686ee
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
