+++ Runtime/Scripts/Core/TypeParser.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 4e270211a5109d14a9eb8ff1f3daea1f
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
