+++ Runtime/Scripts/UI/ConsoleWindow.cs
@@ -13,9 +13,9 @@ namespace DragynGames.Console.UI
         [Header("Visibility")]
         [SerializeField] KeyCode toggleVisabilty = KeyCode.L;
         [SerializeField, Range(0, 1)] float visbleAlpha = 1;
-        
 
-       
+
+
         [Space(10)]
         [Header("Message area")]
         [SerializeField] Transform windowContent;
@@ -46,7 +46,7 @@ namespace DragynGames.Console.UI
             inputField.onValueChanged.AddListener(inputField_OnChanged);
             inputField.onDeselect.AddListener(HandleDeselect);
             inputField.onEndEdit.AddListener(HandleDeselect);
-            MethodHandler.OnSearchComplete += ShowAutocomplete;
+            // MethodHandler.OnSearchComplete += ShowAutocomplete;
         }
 
         private void HandleDeselect(string text)
@@ -69,7 +69,7 @@ namespace DragynGames.Console.UI
                 return;
             }
 
-            MethodHandler.FindMethodsStartingAsync(arg0.TrimStart(commandPrefix));
+            //MethodHandler.FindMethodsStartingAsync(arg0.TrimStart(commandPrefix));
 
         }
 
@@ -85,10 +85,13 @@ namespace DragynGames.Console.UI
 
             if (IsCommand(consoleInput))
             {
-                string response = MethodHandler.ExecuteMethod(consoleInput.TrimStart(commandPrefix));
-                if (!string.IsNullOrEmpty(response))
+                string command = consoleInput.Trim(commandPrefix);
+
+
+                MethodHandler.ExecuteCommand(command);
+                //if (!string.IsNullOrEmpty(response))
                 {
-                    AddMessage(response);
+                    //  AddMessage(response);
                 }
 
             }
@@ -97,7 +100,7 @@ namespace DragynGames.Console.UI
                 AddMessage(consoleInput);
             }
             RemoveTips();
-            
+
         }
 
         private bool IsCommand(string consoleInput)
@@ -118,7 +121,7 @@ namespace DragynGames.Console.UI
         }
         private void Update()
         {
-            if(Input.GetKeyDown(toggleVisabilty))
+            if (Input.GetKeyDown(toggleVisabilty))
             {
                 SetVisability(!visible);
             }
@@ -145,7 +148,7 @@ namespace DragynGames.Console.UI
                 newTip.transform.SetParent(commandTipArea, false);
                 if (i == 5) { break; }
             }
-            int amountOfTime = methodDescriptions.Count > maxNumberOfTips ? maxNumberOfTips:methodDescriptions.Count;
+            int amountOfTime = methodDescriptions.Count > maxNumberOfTips ? maxNumberOfTips : methodDescriptions.Count;
 
             UpdateTipLayout(amountOfTime);
         }
